Site detay sayfaları tamamlandı, ilanlar ile özellikler arasında sonsuza sonsuz ilişki kuruldu ve bu yönde düzenlemeler tamamlandı, Özellikler seçili olarak gelmesi sağlandı

+++ Project_UI/Areas/Admin/Models/ProjectViewModel.cs
@@ -6,7 +6,7 @@ namespace Project_UI.Areas.Admin.Models
     {
         public int IlceId { get; set; }
         public string ProjectFirm { get; set; }
-        public decimal ProjectArea { get; set; }
+        public int ProjectArea { get; set; }
         public int FlatCount { get; set; }
         public DateTime ProjectDeliveryDate { get; set; }
         public string ProjectPromotionVideo { get; set; }
